+++ js/util.js
@@ -1,0 +1,16 @@
+export function getRandomNumber(min, max) {
+  if (max <= min) {
+    [min, max] = [max, min];
+  }
+  if (min <= 0) {
+    min = 0;
+  }
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
+
+
+export function isValidString(string, maxLength) {
+  return string.length <= maxLength;
+}
+
+
